requirements.txt
pynacl
redis
python>=3.7
jsonpickle
gevent
